Cleanup refactoring, adding docstrings
