+++ vantage6/node/docker/docker_manager.py
@@ -4,9 +4,6 @@ The docker manager is responsible for communicating with the docker-
 daemon and is a wrapper arround the docker module. It has methods
 for creating docker networks, docker volumes, start containers
 and retreive results from finished containers
-
-TODO the task folder is also created by this class. This folder needs
-to be cleaned at some point.
 """
 import os
 import time
@@ -15,16 +12,14 @@ import docker
 import re
 import shutil
 
-from typing import Dict, List, NamedTuple
-from docker.models.containers import Container
+from typing import Dict, List, NamedTuple, Union
 from pathlib import Path
 
-from vantage6.common.docker_addons import pull_if_newer
 from vantage6.common.globals import APPNAME
 from vantage6.node.docker.vpn_manager import VPNManager
 from vantage6.node.util import logger_name
 from vantage6.node.docker.network_manager import IsolatedNetworkManager
-from vantage6.node.docker.utils import running_in_docker
+from vantage6.node.docker.task_manager import DockerTaskManager
 
 log = logging.getLogger(logger_name(__name__))
 
@@ -35,258 +30,6 @@ class Result(NamedTuple):
     logs: str
     data: str
     status_code: int
-
-
-class DockerTask(object):
-    log = logging.getLogger(logger_name(__name__))
-
-    def __init__(self, image: str, vpn_manager: VPNManager, node_name: str,
-                 result_id: int, tasks_dir: Path,
-                 isolated_network_mgr: IsolatedNetworkManager,
-                 database_uri: str, database_is_file: bool):
-        # TODO make vars private
-        self.image = image
-        self.vpn_manager = vpn_manager
-        self.node_name = node_name
-        self.result_id = result_id
-        self.__tasks_dir = tasks_dir
-        self.isolated_network_mgr = isolated_network_mgr
-        self.database_uri = database_uri
-        self.database_is_file = database_is_file
-
-        self.docker = docker.from_env()
-        self.container = None
-        self.status_code = None
-
-        self.labels = {
-            f"{APPNAME}-type": "algorithm",
-            "node": self.node_name,
-            "result_id": str(result_id)
-        }
-        self.helper_labels = self.labels
-        self.helper_labels[f"{APPNAME}-type"] = "algorithm-helper"
-
-        # FIXME: these values should be retrieved from DockerNodeContext
-        #   in some way.
-        self.tmp_folder = "/mnt/tmp"
-        self.data_folder = "/mnt/data"
-
-    def is_finished(self):
-        """ Return True if algorithm container is finished """
-        self.container.reload()
-        return self.container.status == 'exited'
-
-    def report_status(self):
-        logs = self.container.logs().decode('utf8')
-
-        # report if the container has a different status than 0
-        self.status_code = self.container.attrs["State"]["ExitCode"]
-        if self.status_code:
-            self.log.error(f"Received non-zero exitcode: {self.status_code}")
-            self.log.error(f"  Container id: {self.container.id}")
-            self.log.warn("Will not remove container")
-            self.log.info(logs)
-        return logs
-
-    def get_results(self):
-        with open(self.output_file, "rb") as fp:
-            results = fp.read()
-        return results
-
-    def pull(self):
-        """Pull the latest image."""
-        try:
-            self.log.info(f"Retrieving latest image: '{self.image}'")
-            pull_if_newer(self.docker, self.image, self.log)
-
-        except Exception as e:
-            self.log.debug('Failed to pull image')
-            self.log.error(e)
-
-    def run_algorithm(self):
-        vpn_port = None
-        if self.vpn_manager.has_vpn:
-            # if VPN is active, network exceptions must be configured
-            # First, start a container that runs indefinitely. The algorithm
-            # container will run in the same network and network exceptions
-            # will therefore also affect the algorithm.
-            self.helper_container = self.docker.containers.run(
-                command='sleep infinity',
-                image='alpine',
-                labels=self.helper_labels,
-                network=self.isolated_network_mgr.network_name,
-                detach=True
-            )
-            # setup forwarding of traffic via VPN client to and from the
-            # algorithm container:
-            vpn_port = self.vpn_manager.forward_vpn_traffic(
-                algo_container=self.helper_container)
-
-        # Try to pull the latest image
-        self.pull()
-
-        # attempt to run the image
-        try:
-            self.log.info(f"Run docker image={self.image}")
-            self.container = self.docker.containers.run(
-                self.image,
-                detach=True,
-                environment=self.environment_variables,
-                network='container:' + self.helper_container.id,
-                volumes=self.volumes,
-                labels=self.labels
-            )
-        except Exception as e:
-            self.log.error('Could not run docker image!?')
-            self.log.error(e)
-            return None
-
-        return vpn_port
-
-    def _make_task_folders(self):
-        # FIXME: We should have a separate mount/volume for this. At the
-        #   moment this is a potential leak as containers might access input,
-        #   output and token from other containers.
-        #
-        #   This was not possible yet as mounting volumes from containers
-        #   is terrible when working from windows (as you have to convert
-        #   from windows to unix several times...).
-
-        # If we're running in docker __tasks_dir will point to a location on
-        # the data volume.
-        # Alternatively, if we're not running in docker it should point to the
-        # folder on the host that can act like a data volume. In both cases,
-        # we can just copy the required files to it
-        self.task_folder_name = f"task-{self.result_id:09d}"
-        self.task_folder_path = \
-            os.path.join(self.__tasks_dir, self.task_folder_name)
-        os.makedirs(self.task_folder_path, exist_ok=True)
-        self.output_file = os.path.join(self.task_folder_path, "output")
-
-    def _prepare_volumes(self, docker_input, tmp_vol_name, token):
-        if isinstance(docker_input, str):
-            docker_input = docker_input.encode('utf8')
-
-        # Create I/O files & token for the algorithm container
-        self.log.debug("prepare IO files in docker volume")
-        io_files = [
-            ('input', docker_input),
-            ('output', b''),
-            ('token', token.encode("ascii")),
-        ]
-
-        for (filename, data) in io_files:
-            filepath = os.path.join(self.task_folder_path, filename)
-
-            with open(filepath, 'wb') as fp:
-                fp.write(data)
-
-        volumes = {
-            tmp_vol_name: {"bind": self.tmp_folder, "mode": "rw"},
-        }
-
-        if running_in_docker():
-            volumes[self.data_volume_name] = \
-                {"bind": self.data_folder, "mode": "rw"}
-        else:
-            volumes[self.__tasks_dir] = \
-                {"bind": self.data_folder, "mode": "rw"}
-        return volumes
-
-    def _setup_environment_vars(self, algorithm_env: Dict = {}):
-        try:
-            proxy_host = os.environ['PROXY_SERVER_HOST']
-
-        except Exception:
-            print('-' * 80)
-            print(os.environ)
-            print('-' * 80)
-            proxy_host = 'host.docker.internal'
-
-        # define enviroment variables for the docker-container, the
-        # host, port and api_path are from the local proxy server to
-        # facilitate indirect communication with the central server
-        # FIXME: we should only prepend data_folder if database_uri is a
-        #   filename
-        environment_variables = {
-            "INPUT_FILE": f"{self.data_folder}/{self.task_folder_name}/input",
-            "OUTPUT_FILE":
-                f"{self.data_folder}/{self.task_folder_name}/output",
-            "TOKEN_FILE": f"{self.data_folder}/{self.task_folder_name}/token",
-            "TEMPORARY_FOLDER": self.tmp_folder,
-            "HOST": f"http://{proxy_host}",
-            "PORT": os.environ.get("PROXY_SERVER_PORT", 8080),
-            "API_PATH": "",
-        }
-        if self.database_is_file:
-            environment_variables["DATABASE_URI"] = \
-                f"{self.data_folder}/{self.database_uri}"
-        else:
-            environment_variables["DATABASE_URI"] = self.database_uri
-        self.log.debug(f"environment: {environment_variables}")
-
-        # Load additional environment variables
-        if algorithm_env:
-            environment_variables = \
-                {**environment_variables, **algorithm_env}
-            self.log.info('Custom environment variables are loaded!')
-            self.log.debug(f"custom environment: {algorithm_env}")
-        return environment_variables
-
-    def run(self, docker_input: bytes, tmp_vol_name: int, token: str,
-            algorithm_env: Dict) -> int:
-        """Runs the docker-image in detached mode.
-
-            It will will attach all mounts (input, output and datafile)
-            to the docker image. And will supply some environment
-            variables.
-
-            :param result_id: server result identifier
-            :param image: docker image name
-            :param docker_input: input that can be read by docker container
-            :param run_id: identifieer of the run sequence
-            :param token: Bearer token that the container can use
-        """
-        # generate task folders
-        self._make_task_folders()
-
-        # prepare volumes
-        self.volumes = self._prepare_volumes(
-            docker_input, tmp_vol_name, token
-        )
-        self.log.debug(f"volumes: {self.volumes}")
-
-        # setup environment variables
-        self.environment_variables = \
-            self._setup_environment_vars(algorithm_env=algorithm_env)
-
-        # run the algorithm as docker container
-        vpn_port = self.run_algorithm()
-        return vpn_port
-
-    def cleanup(self):
-        self._remove_container(self.helper_container, kill=True)
-        if not self.status_code:
-            self._remove_container(self.container)
-
-    def _remove_container(self, container: Container, kill=False) -> None:
-        """
-        Removes a docker container
-
-        Parameters
-        ----------
-        container: docker.models.containers.Container
-            The container that should be removed
-        kill: bool
-            Whether or not container should be killed before it is removed
-        """
-        try:
-            if kill:
-                container.kill()
-            container.remove()
-        except Exception as e:
-            self.log.error(f"Failed to remove container {container.name}")
-            self.log.debug(e)
 
 
 class DockerManager(object):
@@ -300,18 +43,23 @@ class DockerManager(object):
     """
     log = logging.getLogger(logger_name(__name__))
 
-    def __init__(self, ctx, config,
-                 isolated_network_mgr: IsolatedNetworkManager,
+    def __init__(self, ctx, isolated_network_mgr: IsolatedNetworkManager,
                  vpn_manager: VPNManager) -> None:
         """ Initialization of DockerManager creates docker connection and
             sets some default values.
 
-            :param allowed_repositories: allowed urls for docker-images.
-                Empty list implies that all repositoies are allowed.
-            :param tasks_dir: folder to store task related data.
+            Parameters
+            ----------
+            ctx: DockerNodeContext or NodeContext
+                Context object from which some settings are obtained
+            isolated_network_mgr: IsolatedNetworkManager
+                Manager for the isolated network
+            vpn_manager: VPNManager
+                VPN Manager object
         """
         self.log.debug("Initializing DockerManager")
         self.data_volume_name = ctx.docker_volume_name
+        config = ctx.config
         self.algorithm_env = config.get('algorithm_env', {})
         self.vpn_manager = vpn_manager
 
@@ -319,7 +67,7 @@ class DockerManager(object):
         self.docker = docker.from_env()
 
         # keep track of the running containers
-        self.active_tasks: List[DockerTask] = []
+        self.active_tasks: List[DockerTaskManager] = []
 
         # before a task is executed it gets exposed to these regex
         self._allowed_images = config.get("allowed_images")
@@ -345,7 +93,15 @@ class DockerManager(object):
         # set database uri and whether or not it is a file
         self._set_database(config)
 
-    def _set_task_dir(self, ctx):
+    def _set_task_dir(self, ctx) -> None:
+        """
+        Set the task dir
+
+        Parameters
+        ----------
+        ctx: DockerNodeContext or NodeContext
+            Context object containing settings
+        """
         # If we're in a 'regular' context, we'll copy the dataset to our data
         # dir and mount it in any algorithm container that's run; bind mounts
         # on a folder will work just fine.
@@ -375,7 +131,15 @@ class DockerManager(object):
         else:
             self.__tasks_dir = ctx.data_dir
 
-    def _set_database(self, config):
+    def _set_database(self, config: Dict) -> None:
+        """"
+        Set database location and whether or not it is a file
+
+        Parameters
+        ----------
+        config: Dict
+            Configuration of the app
+        """
         # If we're running in a docker container, database_uri would point
         # to a path on the *host* (since it's been read from the config
         # file). That's no good here. Therefore, we expect the CLI to set
@@ -396,14 +160,18 @@ class DockerManager(object):
             self.database_uri = os.path.basename(self.database_uri)
             self.database_is_file = True
 
-    def create_volume(self, volume_name: str):
-        """Create a temporary volume for a single run.
+    def create_volume(self, volume_name: str) -> None:
+        """
+        Create a temporary volume for a single run.
 
-            A single run can consist of multiple algorithm containers.
-            It is important to note that all algorithm containers having
-            the same run_id have access to this container.
+        A single run can consist of multiple algorithm containers. It is
+        important to note that all algorithm containers having the same run_id
+        have access to this container.
 
-            :param run_id: integer representing the run_id
+        Parameters
+        ----------
+        volume_name: str
+            Name of the volume to be created
         """
         try:
             self.docker.volumes.get(volume_name)
@@ -413,14 +181,22 @@ class DockerManager(object):
             self.log.debug(f"Creating volume {volume_name}")
             self.docker.volumes.create(volume_name)
 
-    def is_docker_image_allowed(self, docker_image_name: str):
-        """ Checks the docker image name.
+    def is_docker_image_allowed(self, docker_image_name: str) -> bool:
+        """
+        Checks the docker image name.
 
-            Against a list of regular expressions as defined in the
-            configuration file. If no expressions are defined, all
-            docker images are accepted.
+        Against a list of regular expressions as defined in the configuration
+        file. If no expressions are defined, all docker images are accepted.
 
-            :param docker_image_name: uri to the docker image
+        Parameters
+        ----------
+        docker_image_name: str
+            uri to the docker image
+
+        Returns
+        -------
+        bool
+            Whether docker image is allowed or not
         """
 
         # if no limits are declared
@@ -437,8 +213,20 @@ class DockerManager(object):
         # if not, it is considered an illegal image
         return False
 
-    def is_running(self, result_id):
-        """Return True if a container is already running for <result_id>."""
+    def is_running(self, result_id: int) -> bool:
+        """
+        Check if a container is already running for <result_id>.
+
+        Parameters
+        ----------
+        result_id: int
+            result_id of the algorithm container to be found
+
+        Returns
+        -------
+        bool
+            Whether or not algorithm container is running already
+        """
         running_containers = self.docker.containers.list(filters={
             "label": [
                 f"{APPNAME}-type=algorithm",
@@ -449,18 +237,28 @@ class DockerManager(object):
         return bool(running_containers)
 
     def run(self, result_id: int,  image: str, docker_input: bytes,
-            tmp_vol_name: int, token: str) -> int:
-        """Runs the docker-image in detached mode.
+            tmp_vol_name: str, token: str) -> Union[int, None]:
+        """
+        Checks if docker task is running. If not, creates DockerTaskManager to
+        run the task
 
-            It will will attach all mounts (input, output and datafile)
-            to the docker image. And will supply some environment
-            variables.
+        Parameters
+        ----------
+        result_id: int
+            Server result identifier
+        image: str
+            Docker image name
+        docker_input: bytes
+            Input that can be read by docker container
+        tmp_vol_name: str
+            Name of temporary docker volume assigned to the algorithm
+        token: str
+            Bearer token that the container can use
 
-            :param result_id: server result identifier
-            :param image: docker image name
-            :param docker_input: input that can be read by docker container
-            :param run_id: identifieer of the run sequence
-            :param token: Bearer token that the container can use
+        Returns
+        -------
+        int or None:
+            Port number assigned for VPN communication. None if VPN is inactive
         """
         # Verify that an allowed image is used
         if not self.is_docker_image_allowed(image):
@@ -474,7 +272,7 @@ class DockerManager(object):
             self.log.debug(f"result_id={result_id} is discarded")
             return None
 
-        task = DockerTask(
+        task = DockerTaskManager(
             image=image,
             result_id=result_id,
             vpn_manager=self.vpn_manager,
@@ -494,7 +292,7 @@ class DockerManager(object):
 
         return vpn_port
 
-    def get_result(self):
+    def get_result(self) -> Result:
         """
         Returns the oldest (FIFO) finished docker container.
 
@@ -538,8 +336,16 @@ class DockerManager(object):
             status_code=finished_task.status_code
         )
 
-    def login_to_registries(self, registies: list = []) -> None:
-        for registry in registies:
+    def login_to_registries(self, registries: list = []) -> None:
+        """
+        Login to the docker registries
+
+        Parameters
+        ----------
+        registries: list
+            list of registries to login to
+        """
+        for registry in registries:
             try:
                 self.docker.login(
                     username=registry.get("username"),
